// src/modules/notifications/enums/notification-type.enum.ts
export enum NotificationType {
  SUCCESSFUL_REGISTRATION = 'SUCCESSFUL_REGISTRATION',
  WORKER_ACCOUNT_VERIFICATION = 'WORKER_ACCOUNT_VERIFICATION',
  NEW_BOOKING_REQUEST = 'NEW_BOOKING_REQUEST',
  BOOKING_ACCEPTED = 'BOOKING_ACCEPTED',
  BOOKING_REJECTED = 'BOOKING_REJECTED',
  BOOKING_REMINDER = 'BOOKING_REMINDER',
  PAYMENT_CONFIRMATION = 'PAYMENT_CONFIRMATION',
  SERVICE_COMPLETION = 'SERVICE_COMPLETION',
  COMPLAINT_STATUS_UPDATE = 'COMPLAINT_STATUS_UPDATE',
  CUSTOMER_REVIEW = 'CUSTOMER_REVIEW',
  SYSTEM_ANNOUNCEMENT = 'SYSTEM_ANNOUNCEMENT',
  PROMOTIONAL_ANNOUNCEMENT = 'PROMOTIONAL_ANNOUNCEMENT',
}